+++ createDirectoriesTree.sh
@@ -7,9 +7,7 @@ cd ~/ &&
 mkdir .gradle
 mkdir Coding
 cd Coding &&
-mkdir IDEs
 mkdir xDKs
-mkdir servers
 mkdir tools
 mkdir workspace-mine
 mkdir workspace-test